requirement.txt
fyers-apiv2
python-dotenv